fix: Major refactoring, splitting code into separate files depending on their areas: "rundown": General rundown manipulation, "media-scanner": Media-scanner, "mos": mos-specific functions
