+++ meteor/server/api/integration/media-scanner.ts
@@ -1,0 +1,80 @@
+import { Meteor } from 'meteor/meteor'
+import { check, Match } from 'meteor/check'
+import * as _ from 'underscore'
+import { PeripheralDeviceAPI } from '../../../lib/api/peripheralDevice'
+import { PeripheralDevices } from '../../../lib/collections/PeripheralDevices'
+import { RunningOrder, RunningOrders, DBRunningOrder } from '../../../lib/collections/RunningOrders'
+import { SegmentLine, SegmentLines, DBSegmentLine, SegmentLineHoldMode, SegmentLineNoteType, SegmentLineNote } from '../../../lib/collections/SegmentLines'
+import { SegmentLineItem, SegmentLineItems } from '../../../lib/collections/SegmentLineItems'
+import { Segments, DBSegment, Segment } from '../../../lib/collections/Segments'
+import { saveIntoDb, partialExceptId, getCurrentTime, literal, fetchBefore, getRank, fetchAfter } from '../../../lib/lib'
+import { PeripheralDeviceSecurity } from '../../security/peripheralDevices'
+import { PeripheralDeviceCommands } from '../../../lib/collections/PeripheralDeviceCommands'
+import { logger } from '../../logging'
+import { runTemplate, runNamedTemplate, TemplateContext, RunTemplateResult, TemplateResultAfterPost } from '../templates/templates'
+import { getHash } from '../../lib'
+import { Timeline } from '../../../lib/collections/Timeline'
+import { StudioInstallations, StudioInstallation } from '../../../lib/collections/StudioInstallations'
+import { MediaObject, MediaObjects } from '../../../lib/collections/MediaObjects'
+import { SegmentLineAdLibItem, SegmentLineAdLibItems } from '../../../lib/collections/SegmentLineAdLibItems'
+import { ShowStyles, ShowStyle } from '../../../lib/collections/ShowStyles'
+import { ServerPlayoutAPI, updateTimelineFromMosData, updateTimeline, afterUpdateTimeline } from '../playout'
+import { syncFunction } from '../../codeControl'
+import { CachePrefix } from '../../../lib/collections/RunningOrderDataCache'
+import { setMeteorMethods, Methods, wrapMethods } from '../../methods'
+
+export namespace MediaScannerIntegration {
+	export function getMediaObjectRevisions (id, token, collectionId: string) {
+		logger.debug('getMediaObjectRevisions')
+		let peripheralDevice = PeripheralDeviceSecurity.getPeripheralDevice(id, token, this)
+
+		if (peripheralDevice.studioInstallationId) {
+			return _.map(MediaObjects.find({
+				studioId: peripheralDevice.studioInstallationId,
+				collectionId: collectionId
+			}).fetch(), (mo: MediaObject) => {
+				return {
+					id: 	mo.objId,
+					rev: 	mo._rev
+				}
+			})
+		} else {
+			throw new Meteor.Error(400, 'Device "' + peripheralDevice._id + '" has no studioInstallation')
+		}
+	}
+	export function updateMediaObject (id, token, collectionId: string, objId, doc: MediaObject | null) {
+		logger.debug('updateMediaObject')
+		let peripheralDevice = PeripheralDeviceSecurity.getPeripheralDevice(id, token, this)
+
+		let _id = collectionId + '_' + objId
+		if (_.isNull(doc)) {
+			MediaObjects.remove(_id)
+		} else if (doc) {
+			let doc2 = _.extend(doc, {
+				studioId: peripheralDevice.studioInstallationId,
+				collectionId: collectionId,
+				objId: objId,
+				_id: _id
+			})
+			// logger.debug(doc2)
+			MediaObjects.upsert(_id, {$set: doc2})
+		} else {
+			throw new Meteor.Error(400, 'missing doc argument')
+		}
+	}
+	export const executeFunction: (deviceId: string, functionName: string, ...args: any[]) => any = Meteor.wrapAsync((deviceId: string, functionName: string, ...args: any[]) => {
+		let args0 = args.slice(0, -1)
+		let cb = args.slice(-1)[0] // the last argument in ...args
+		PeripheralDeviceAPI.executeFunction(deviceId, cb, functionName, ...args0)
+	})
+}
+
+let methods: Methods = {}
+methods[PeripheralDeviceAPI.methods.getMediaObjectRevisions] = (deviceId, deviceToken, collectionId: string,) => {
+	return MediaScannerIntegration.getMediaObjectRevisions(deviceId, deviceToken, collectionId)
+}
+methods[PeripheralDeviceAPI.methods.updateMediaObject] = (deviceId, deviceToken, collectionId: string, id: string, doc: MediaObject | null) => {
+	return MediaScannerIntegration.updateMediaObject(deviceId, deviceToken, collectionId, id, doc)
+}
+// Apply methods:
+setMeteorMethods(wrapMethods(methods))
